Migrate user-service to TypeScript
