+++ services/admin-service/openapiDoc.json
@@ -7,7 +7,7 @@
   },
   "servers": [
     {
-      "url": "http://localhost:4000/"
+      "url": "http://localhost:5005/"
     }
   ],
   "paths": {
