+++ services/admin-service/package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start:dev": "yarn nodemon",
-    "swagger-autogen": "ts-node ./openapiDocGen.ts"
+    "start:dev": "yarn swagger-autogen && ts-node-dev ./app.ts",
+    "swagger-autogen": "ts-node-dev ./openapiDocGen.ts"
   },
   "dependencies": {
     "dotenv": "^16.3.1",
@@ -14,9 +14,8 @@
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/node": "^20.6.2",
-    "nodemon": "^3.0.1",
     "swagger-autogen": "^2.23.6",
-    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
     "typescript": "^5.2.2"
   }
 }
