--- services/user-service/db/prismaClient.js
@@ -1,5 +1,0 @@
-const {PrismaClient} = require('@prisma/client');
-
-const prismaClient = new PrismaClient();
-
-module.exports=prismaClient
