+++ services/user-service/db/prismaClient.ts
@@ -1,0 +1,5 @@
+import {PrismaClient} from '@prisma/client';
+
+const prismaClient : PrismaClient = new PrismaClient();
+
+export default prismaClient;
