+++ services/user-service/package.json
@@ -3,22 +3,22 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start:dev": "yarn nodemon",
+    "start:dev": "yarn swagger-autogen && ts-node-dev ./app.ts",
+    "swagger-autogen": "ts-node-dev ./openapiDocGen.ts",
     "build": "tsc"
   },
   "dependencies": {
     "debug": "~2.6.9",
     "express": "^4.18.2",
-    "express-openapi": "^12.1.3",
     "morgan": "~1.9.1"
   },
   "devDependencies": {
-    "nodemon": "^3.0.1",
     "@types/cookie-parser": "^1.4.4",
     "@types/cors": "^2.8.14",
     "@types/express": "^4.17.17",
     "@types/morgan": "^1.9.5",
     "@types/node": "^20.6.2",
+    "swagger-autogen": "^2.23.6",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.2.2"
   }
